Refactor run and stop command handling
